Move processingGuarantee up from Processor.Context
Method `processingGuarantee()` was available only in `Processor.Context`,
but this information is available up in `ProcessorMetaSupplier.Context`,
it should be there.
